Stable codebase at refactor
Instead of having an 'api' Django app, have two apps 'notes_api' and 'todos_api'
