--- api/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import Note, Todo
-
-
-admin.site.register(Note)
-admin.site.register(Todo)
