+++ notes_api/apps.py
@@ -3,5 +3,5 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 
-class ApiConfig(AppConfig):
-    name = 'api'
+class NotesApiConfig(AppConfig):
+    name = 'notes_api'
